// src/enums/message-topic.ts
enum MessageTopic {
    JOIN_GAME = 'join.game',
    WELCOME = 'welcome',
    READY = 'ready',
    SET_WINNER = 'set.winner',
    NEXT_ROUND = 'next.round',
    USER_JOINED = 'player.joined',
    ROUND_FINISHED = 'round.finished',
    RANKING = 'ranking',
    GAME_FINISHED = 'game.finished'
}

export default MessageTopic